初版迁移

+++ shared/models/nonsense.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, Integer, Text
+from sqlalchemy import BigInteger, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from shared.models.base import Base, TimestampMixin
@@ -11,5 +11,4 @@ class Nonsense(Base, TimestampMixin):
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     content: Mapped[str] = mapped_column(Text, nullable=False, default="")
-    send_times: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-
+    send_times: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
